backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
pydantic-settings
pwdlib[bcrypt]
argon2-cffi
python-jose
python-multipart